refactor(core): extract maintenance/status.ts + shared with-db helper
Second sub-commit of the maintenance/ split (codemem-ug38). Moves two
raw-event functions out of maintenance.ts verbatim:

 - getRawEventStatus  — raw-event backlog status read-side
 - retryRawEventFailures — requeues failed/errored flush batches

Both are self-contained, share the withDb wrapper, and have no
cross-dependency on memory-role / relink / dedup logic still in
maintenance.ts.

Also extracts the withDb helper itself into a small maintenance/with-db.ts
module so future job-family extractions can share it without each one
either duplicating the body or importing from maintenance.ts (which
would invert the dependency direction).

maintenance.ts re-exports getRawEventStatus + retryRawEventFailures
from the new status module so external callers (memory-role CLI, health
tab, tests) import from the same place.

Mechanical copy — function bodies byte-for-byte, withDb body byte-for-byte.
Gate stays green (tsc / lint / 1417 tests). maintenance.ts drops from
2602 to 2458 lines.

Left out intentionally this round:
 - rawEventsGate — depends on getReliabilityMetrics which hasn't been
   extracted yet; goes with the reliability.ts slice.
